+++ CHANGES.rst
@@ -19,6 +19,8 @@ Improvements
   whole browser when there are thousands of log entries
 - Add shortcut to next upcoming event in a category (:issue:`3388`)
 - Make registration period display less confusing (:issue:`3359`)
+- Add setting for the default contribution duration of an event
+  (:issue:`3446`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^